Part 13 ToDo continue implementing Android Page Object for the different functionality - Calendar, Select Adults Number...
Passing first few steps until Calendar popup in Android, Web passing all.
At that point part 13 my level of completetion – the AUT is changed with new functionality for Calendar and I will need to adjust my framework to respond to the change. I need to implement the Page Object Classes in the Android package with the new locators. And record a demo of Appium Server, Eclipse, Emulator of the test being cross platform – first chrome than android. – 1 minute video and add it in the README.md in GitHub.